+++ Sources/RichTextKit/Data/RichTextDataError.swift
@@ -3,17 +3,16 @@
 //  RichTextKit
 //
 //  Created by Daniel Saidi on 2022-06-04.
-//  Copyright © 2022-2023 Daniel Saidi. All rights reserved.
+//  Copyright © 2022-2024 Daniel Saidi. All rights reserved.
 //
 
 import Foundation
 
 /**
- This enum represents errors that can be thrown when getting
- rich text data for a certain format.
+ This enum represents rich text data-related errors.
  */
-public enum RichTextDataError: Error {
-
+enum RichTextDataError: Error {
+    
     case invalidArchivedData(in: Data)
     case invalidPlainTextData(in: Data)
     case invalidData(in: String)