+++ Sources/RichTextKit/_Deprecated/RichTextData+Deprecated.swift
@@ -1,0 +1,4 @@
+import SwiftUI
+
+@available(*, deprecated, renamed: "RichTextDataFormat.Menu")
+public typealias RichTextDataFormatMenu = RichTextDataFormat.Menu